Move commonly required classes into separate module
