+++ common/__init__.py
@@ -1,0 +1,6 @@
+from .ptconfig import load_config, write_config
+from .resolver import Resolver
+
+__all__ = [
+    load_config, write_config
+]
